Revert "Move prototype files into correct folders"
This reverts commit 1b9e58331969ebfae750be21ef1a6374584ea6a0.
