+++ scripts/demo_simplified.py
@@ -24,7 +24,7 @@ import torch
 PROJECT_ROOT = Path(__file__).parent.parent
 sys.path.insert(0, str(PROJECT_ROOT))
 
-from hrm.prototype.models.model_simplified import SimplifiedHRM, SimplifiedHRMConfig, PuzzleType
+from hrm.model_simplified import SimplifiedHRM, SimplifiedHRMConfig, PuzzleType
 
 
 def load_model(model_path: str, device: torch.device) -> SimplifiedHRM:
