+++ scripts/train_simplified.py
@@ -51,7 +51,7 @@ from torch.utils.data import DataLoader, Dataset
 PROJECT_ROOT = Path(__file__).parent.parent
 sys.path.insert(0, str(PROJECT_ROOT))
 
-from hrm.prototype.models.model_simplified import (
+from hrm.model_simplified import (
     SimplifiedHRM,
     SimplifiedHRMConfig,
     PuzzleType,
